Unnecessary casts
git-svn-id: https://svn.apache.org/repos/asf/commons/proper/sanselan/trunk@922388 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/java/org/apache/sanselan/common/RationalNumber.java
@@ -91,7 +91,7 @@ public class RationalNumber extends Number
 
 	public int intValue()
 	{
-		return (int) numerator / (int) divisor;
+		return numerator / divisor;
 	}
 
 	public long longValue()
